refactor package of AboutUsFragment
